Refactor tests into a new file
